# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Universe Moon [Um] API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /photos:
    get:
      operationId: getPhotos
      tags: [gallery]
      summary: Get photos
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Photo"
    post:
      operationId: addPhoto
      tags: [gallery]
      summary: Add photo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPhotoRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
  /photos/{id}:
    delete:
      operationId: deletePhoto
      tags: [gallery]
      summary: Delete photo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /memories:
    get:
      operationId: getMemories
      tags: [memories]
      summary: Get memories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memory"
    post:
      operationId: addMemory
      tags: [memories]
      summary: Add memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMemoryRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Memory"
  /memories/{id}:
    delete:
      operationId: deleteMemory
      tags: [memories]
      summary: Delete memory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /links:
    get:
      operationId: getLinks
      tags: [links]
      summary: Get links
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Link"
    post:
      operationId: addLink
      tags: [links]
      summary: Add link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddLinkRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Link"
  /links/{id}:
    delete:
      operationId: deleteLink
      tags: [links]
      summary: Delete link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /music:
    get:
      operationId: getMusic
      tags: [music]
      summary: Get music playlist
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
    post:
      operationId: addTrack
      tags: [music]
      summary: Add track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTrackRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
  /music/{id}:
    delete:
      operationId: deleteTrack
      tags: [music]
      summary: Delete track
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /secrets:
    get:
      operationId: getSecrets
      tags: [secrets]
      summary: Get vault secrets (requires PIN)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Secret"
    post:
      operationId: addSecret
      tags: [secrets]
      summary: Add secret
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddSecretRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Secret"
  /secrets/{id}:
    put:
      operationId: toggleSecret
      tags: [secrets]
      summary: Toggle secret locked state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Secret"
    delete:
      operationId: deleteSecret
      tags: [secrets]
      summary: Delete secret
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /opmem:
    get:
      operationId: getOpMem
      tags: [opmem]
      summary: Get OpMem info
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpMem"
    post:
      operationId: createOpMem
      tags: [opmem]
      summary: Create/update OpMem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOpMemRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpMem"
  /opmem/accepted:
    post:
      operationId: addAcceptedMember
      tags: [opmem]
      summary: Add accepted member to OpMem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                opMemId:
                  type: integer
              required: [name, opMemId]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /mep:
    get:
      operationId: getMeps
      tags: [mep]
      summary: Get MEP list
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mep"
    post:
      operationId: createMep
      tags: [mep]
      summary: Create MEP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMepRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mep"
  /mep/{id}:
    delete:
      operationId: deleteMep
      tags: [mep]
      summary: Delete MEP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Photo:
      type: object
      properties:
        id:
          type: integer
        url:
          type: string
        caption:
          type: string
        uploadedBy:
          type: string
        category:
          type: string
        createdAt:
          type: string
      required: [id, url, createdAt]
    AddPhotoRequest:
      type: object
      properties:
        url:
          type: string
        caption:
          type: string
        uploadedBy:
          type: string
        category:
          type: string
      required: [url]
    Memory:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        date:
          type: string
        addedBy:
          type: string
        photo:
          type: string
        createdAt:
          type: string
      required: [id, title, content, createdAt]
    AddMemoryRequest:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        date:
          type: string
        addedBy:
          type: string
        photo:
          type: string
      required: [title, content]
    Link:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        url:
          type: string
        desc:
          type: string
        category:
          type: string
        addedBy:
          type: string
        createdAt:
          type: string
      required: [id, title, url, createdAt]
    AddLinkRequest:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        desc:
          type: string
        category:
          type: string
        addedBy:
          type: string
      required: [title, url]
    Track:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        artist:
          type: string
        url:
          type: string
        addedBy:
          type: string
        createdAt:
          type: string
      required: [id, title, artist, createdAt]
    AddTrackRequest:
      type: object
      properties:
        title:
          type: string
        artist:
          type: string
        url:
          type: string
        addedBy:
          type: string
      required: [title, artist]
    Secret:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        revealDate:
          type: string
        locked:
          type: boolean
        createdAt:
          type: string
      required: [id, content, locked, createdAt]
    AddSecretRequest:
      type: object
      properties:
        content:
          type: string
        revealDate:
          type: string
      required: [content]
    OpMem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        openDate:
          type: string
        closeDate:
          type: string
        isOpen:
          type: boolean
        acceptedMembers:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required: [id, title, isOpen, createdAt]
    CreateOpMemRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        openDate:
          type: string
        closeDate:
          type: string
        isOpen:
          type: boolean
      required: [title]
    Mep:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        participants:
          type: array
          items:
            type: string
        createdBy:
          type: string
        createdAt:
          type: string
      required: [id, title, videoUrl, createdAt]
    CreateMepRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        participants:
          type: array
          items:
            type: string
        createdBy:
          type: string
      required: [title, videoUrl]
...