# lib/api-zod/src/generated/types/addTrackRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Universe Moon [Um] API
 * OpenAPI spec version: 0.1.0
 */

export interface AddTrackRequest {
  title: string;
  artist: string;
  url?: string;
  addedBy?: string;
}