# lib/api-zod/src/generated/types/mep.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Universe Moon [Um] API
 * OpenAPI spec version: 0.1.0
 */

export interface Mep {
  id: number;
  title: string;
  description?: string;
  videoUrl: string;
  participants?: string[];
  createdBy?: string;
  createdAt: string;
}